Adjusted GUI components for features in anticipation of pubs:
1)cropping of new feature names to fit in popupmenus

2)changed button and popup quicknote

Other adjustments:

1) fixed file saving default directories (Issue #34)

2) Fixed gray LUT-still need to add support for more LUTs in scatter plot.
Also added separate LUT for categorical data, ex. less than 10 values.
The beginnings of a VTEA LUT service are sketched out, need to consider
adapting existing imagej2 infrasctructure. 

3) Morphology interface is being fleshed-out, SciJava infrastructure working

4) Gui is not limited to 5 channels (#13)

5) Fixed the 3D button nullpointerexception bug, This was associated with
the gate arraylist not generating the mask imageplus.  This is now forced
prior to making the 3D volume. (#6) 

6) Closed overlay images etc warnings streamlined to make less annoying.

7) Multiple segmentations can be run in parallel with new Processor approach.

8) Grow Morphology implemented and validation is in progress. Initial tests
show consistent results.  This needs to be formalized with Andre's work on 
in silico ground-truth.