To address issue #81:
Added cluster number estimation by the Elbow method to the Feature 
Menu.  Currently the method defaults to running K-means 100 times for k
from 2 to 30 and reports distortion as calculated by K-means in the 
smile library. Graphs are JfreeChart derived and can be printed and
saved.

Add to beginning of feature list was confusing parallel asynchronous
additions of features.  Corrected to ensure offset in Feature Frame. 
Also fixed the plotting behavior on new features, sets x and y 
logically.